- Fixed sign-overflows during compares of guint32 mac timestamps
- Added function for generation of playlist containing all songs
listened to since the last sync.

git-svn-id: https://gtkpod.svn.sf.net/svnroot/gtkpod/gtkpod/trunk@249 f01d2545-417e-4e96-918e-98f8d0dbbcb6

+++ src/display_tracks.c
@@ -1,4 +1,4 @@
-/* Time-stamp: <2003-08-09 23:54:52 jcs>
+/* Time-stamp: <2003-08-22 22:59:39 jcs>
 |
 |  Copyright (C) 2002-2003 Jorg Schuler <jcsjcs at users.sourceforge.net>
 |  Part of the gtkpod project.
@@ -808,7 +808,8 @@ gint sm_data_compare_func (GtkTreeModel *model,
       return song1->rating - song2->rating;
   case SM_COLUMN_TIME_PLAYED:
   case SM_COLUMN_TIME_MODIFIED:
-      return time_get_time (song1, sm_item) - time_get_time (song2, sm_item);
+      return ((float)time_get_time (song1, sm_item) -
+	      time_get_time (song2, sm_item));
   case  SM_COLUMN_VOLUME:
       return song1->volume - song2->volume;
   default:
